Terraform runner should generate it's own configs
